# openapi.json
{"openapi": "3.1.0", "info": {"title": "Timer API", "description": "A simple API to set and get timers", "version": "0.1.0"}, "paths": {"/": {"get": {"summary": "Root", "operationId": "root__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/timer": {"post": {"tags": ["timer"], "summary": "Set Timer", "operationId": "set_timer_timer_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/SetTimerRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ApiResponse_Any_Any_"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/timer/{timer_id}": {"get": {"tags": ["timer"], "summary": "Get Timer", "operationId": "get_timer_timer__timer_id__get", "parameters": [{"name": "timer_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Timer Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ApiResponse_GetTimerResponse_Any_"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"ApiResponse_Any_Any_": {"properties": {"errors": {"items": {}, "type": "array", "title": "Errors"}, "data": {"items": {}, "type": "array", "title": "Data"}}, "type": "object", "title": "ApiResponse[Any, Any]"}, "ApiResponse_GetTimerResponse_Any_": {"properties": {"errors": {"items": {}, "type": "array", "title": "Errors"}, "data": {"items": {"$ref": "#/components/schemas/GetTimerResponse"}, "type": "array", "title": "Data"}}, "type": "object", "title": "ApiResponse[GetTimerResponse, Any]"}, "GetTimerResponse": {"properties": {"id": {"type": "string", "title": "Id"}, "seconds_left": {"type": "integer", "title": "Seconds Left"}}, "type": "object", "required": ["id", "seconds_left"], "title": "GetTimerResponse"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "SetTimerRequest": {"properties": {"hours": {"type": "integer", "minimum": 0.0, "title": "Hours"}, "minutes": {"type": "integer", "maximum": 59.0, "minimum": 0.0, "title": "Minutes"}, "seconds": {"type": "integer", "maximum": 59.0, "minimum": 0.0, "title": "Seconds"}, "url": {"type": "string", "title": "Url"}}, "type": "object", "required": ["hours", "minutes", "seconds", "url"], "title": "SetTimerRequest"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}, "tags": [{"name": "timer", "description": "Operations related to timers"}]}